Function xed_sys::xed_interface::xed_format_context

pub unsafe extern "C" fn xed_format_context(
    syntax: xed_syntax_enum_t, 
    xedd: *const xed_decoded_inst_t, 
    out_buffer: *mut c_char, 
    buffer_len: c_int, 
    runtime_instruction_address: u64, 
    context: *mut c_void, 
    symbolic_callback: xed_disassembly_callback_fn_t
) -> xed_bool_t

Disassemble the decoded instruction using the specified syntax. The output buffer must be at least 25 bytes long. Returns true if disassembly proceeded without errors. @param syntax a #xed_syntax_enum_t the specifies the disassembly format @param xedd a #xed_decoded_inst_t for a decoded instruction @param out_buffer a buffer to write the disassembly in to. @param buffer_len maximum length of the disassembly buffer @param runtime_instruction_address the address of the instruction being disassembled. If zero, the offset is printed for relative branches. If nonzero, XED attempts to print the target address for relative branches. @param context A void* used only for the call back routine for symbolic disassembly if one is provided. Can be zero. @param symbolic_callback A function pointer for obtaining symbolic disassembly. Can be zero. @return Returns 0 if the disassembly fails, 1 otherwise. @ingroup PRINT