Struct xed_sys::xed_interface::xed_decoded_inst_s

#[repr(C)]

pub struct xed_decoded_inst_s {
    pub _operands: xed_operand_storage_t,
    pub _operand_order: [u8; 5],
    pub _n_operand_order: u8,
    pub _decoded_length: u8,
    pub _inst: *const xed_inst_t,
    pub _byte_array: xed_decoded_inst_s__bindgen_ty_1,
    pub u: xed_decoded_inst_s__bindgen_ty_2,
}

@ingroup DEC The main container for instructions. After decode, it holds an array of operands with derived information from decode and also valid

xed_inst_t pointer which describes the operand templates and the

operand order. See @ref DEC for API documentation.

Fields
_operands: xed_operand_storage_t

The _operands are storage for information discovered during decoding. They are also used by encode. The accessors for these operands all have the form xed3_operand_{get,set}_*(). They should be considered internal and subject to change over time. It is preferred that you use xed_decoded_inst_*() or the xed_operand_values_*() functions when available.

_operand_order: [u8; 5]

Used for encode operand ordering. Not set by decode.

_n_operand_order: u8

Length of the _operand_order[] array.

_decoded_length: u8 _inst: *const xed_inst_t

when we decode an instruction, we set the _inst and get the properites of that instruction here. This also points to the operands template array.

_byte_array: xed_decoded_inst_s__bindgen_ty_1 u: xed_decoded_inst_s__bindgen_ty_2