Struct xed_sys::xed_interface::xed_print_info_t

#[repr(C)]

pub struct xed_print_info_t {
    pub p: *const xed_decoded_inst_t,
    pub buf: *mut c_char,
    pub blen: c_int,
    pub runtime_address: u64,
    pub disassembly_callback: xed_disassembly_callback_fn_t,
    pub context: *mut c_void,
    pub syntax: xed_syntax_enum_t,
    pub format_options_valid: c_int,
    pub format_options: xed_format_options_t,
    pub emitted: xed_bool_t,
    pub operand_indx: c_uint,
    pub skip_operand: c_uint,
    pub extra_index_operand: xed_reg_enum_t,
    pub implicit: xed_bool_t,
}

@ingroup PRINT This contains the information used by the various disassembly printers. Call xed_init_print_info to initialize the fields. Then change the required and optional fields when required.

Fields
p: *const xed_decoded_inst_t

the decoded instruction to print

buf: *mut c_char

pointer to the output buffer

blen: c_int

length of the output buffer. (bytes) Must be > 25 to start.

runtime_address: u64

program counter location. Must be zero if not used. (Sometimes instructions are disassembled in a temporary buffer at a different location than where they may or will exist in memory).

disassembly_callback: xed_disassembly_callback_fn_t

disassembly_callback MUST be set to zero if not used! If zero, the default disassembly callback is used (if one has been registered).

context: *mut c_void

passed to disassembly callback. Can be zero if not used.

syntax: xed_syntax_enum_t

default is Intel-syntax (dest on left)

format_options_valid: c_int

1=indicated the format_options field is valid, 0=use default formatting options from xed_format_set_options().

format_options: xed_format_options_t emitted: xed_bool_t

internal, do not use

operand_indx: c_uint

internal, do not use

skip_operand: c_uint

internal, do not use

extra_index_operand: xed_reg_enum_t

internal, do not use

implicit: xed_bool_t

internal, do not use