Type Definition xed_sys::xed_interface::xed_disassembly_callback_fn_t

type xed_disassembly_callback_fn_t = Option<unsafe extern "C" fn(_: u64, _: *mut c_char, _: u32, _: *mut u64, _: *mut c_void) -> c_int>;

@param address The input address for which we want symbolic name and offset @param symbol_buffer A buffer to hold the symbol name. The callback function should fill this in and terminate with a null byte. @param buffer_length The maximum length of the symbol_buffer including then null @param offset A pointer to a xed_uint64_t to hold the offset from the provided symbol. @param context This void* pointer passed to the disassembler's new interface so that the caller can identify the proper context against which to resolve the symbols. The disassembler passes this value to the callback. The legacy formatters that do not have context will pass zero for this parameter. @return 0 on failure, 1 on success.